# docs/swagger.yaml
basePath: /ih/mobile_central_2_0_0
definitions:
...
  dto.CreateUserBillingDto:
    properties:
      userBilling:
        $ref: '#/definitions/dto.UserBillingDto'
      userBillingPayment:
        $ref: '#/definitions/dto.UserBillingPaymentDto'
    required:
    - userBilling
    - userBillingPayment
    type: object
...
  dto.UserBillingDto:
    properties:
      patientDocNumber:
        type: string
      patientName:
        type: string
      patientPrn:
        type: string
      vesBillNumber:
        type: string
      vesInvoiceAmount:
        type: string
      vesInvoiceDateTime:
        type: string
      vesInvoiceNumber:
        type: string
      vesOutstandingAmount:
        type: string
      vesRegDateTime:
        type: string
    required:
    - patientDocNumber
    - patientName
    - patientPrn
    - vesBillNumber
    - vesInvoiceAmount
    - vesInvoiceDateTime
    - vesInvoiceNumber
    - vesOutstandingAmount
    - vesRegDateTime
    type: object
  dto.UserBillingPaymentDto:
    properties:
      billingAddress1:
        type: string
      billingAddress2:
        type: string
      billingAddress3:
        type: string
      billingContactCode:
        type: string
      billingContactNo:
        type: string
      billingCountryCode:
        type: string
      billingEmail:
        type: string
      billingFullname:
        type: string
      billingPostcode:
        type: string
      billingState:
        type: string
      billingTowncity:
        type: string
    required:
    - billingAddress1
    - billingContactCode
    - billingContactNo
    - billingCountryCode
    - billingEmail
    - billingFullname
    type: object
  dto.UserBillingPaymentStatusDto:
    properties:
      status:
        type: string
    required:
    - status
    type: object
...
  userBilling.UserBillingPayment:
    properties:
      billingFullname:
        type: string
      patientDocNumber:
        type: string
      patientName:
        type: string
      patientPrn:
        type: string
      paymentAmountCollected:
        type: string
      paymentGateway:
        type: string
      paymentRequestNo:
        type: string
      paymentTransDate:
        type: string
      vesBillNumber:
        type: string
      vesInvoiceAmount:
        type: string
      vesInvoiceDateTime:
        type: string
      vesOutstandingAmount:
        type: string
      vesPaymentReceiptNo:
        type: string
    type: object
...
info:
  contact: {}
  description: Vesalius-m Backend API Docs (IHP-UAT Version).
  title: (IHP-UAT) Vesalius-m Backend API
  version: "1.0"
paths:
...
  /user-billing/paid/all/mobile:
    get:
      parameters:
      - description: _page
        in: query
        name: _page
        type: string
      - description: _limit
        in: query
        name: _limit
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/userBilling.UserBillingPayment'
            type: array
      security:
      - BearerAuth: []
      tags:
      - User Billing
  /user-billing/pay/{paymentMethod}:
    post:
      parameters:
      - description: paymentMethod
        in: path
        name: paymentMethod
        required: true
        type: integer
      - description: CreateUserBillingDto
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateUserBillingDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - BearerAuth: []
      tags:
      - User Billing
  /user-billing/status/{billPaymentId}:
    post:
      parameters:
      - description: billPaymentId
        in: path
        name: billPaymentId
        required: true
        type: string
      - description: UserBillingPaymentStatusDto
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UserBillingPaymentStatusDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - BearerAuth: []
      tags:
      - User Billing
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"